shows earnings of each crew memeber

+++ Hacker.cs
@@ -6,7 +6,7 @@ namespace Heist_II
     {
         public string Name { get; set; }
         public int SkillLevel { get; set; }
-        public int PercentageCut { get; set; }
+        public double PercentageCut { get; set; }
 
         public void Specialty()
         {
